--- src/BooksList.js
@@ -1,20 +1,0 @@
-import React from 'react';
-
-function BooksList() {
-  return (
-    <table>
-      <tr>
-        <th>BookID</th>
-        <th>Title</th>
-        <th>Category</th>
-      </tr>
-      <tr>
-        <td />
-        <td />
-        <td />
-      </tr>
-    </table>
-  );
-}
-
-export default BooksList;
